LifecycleTest - Allow performing API calls via CLI or PHP
Most of the test steps and assertionsa are based on calling the API.  Is it
better for these calls to run to PHP functions (`civicrm_api()`) in the same
process - or is it better to run each in a separate subprocess (`cv api`)?

Both approaches seem useful:

* Running via subprocess is a truer representation of how most interaction happens during sysadmin
* Running via local PHP should help to capture subtle leaks that can manifest as funny interactions during enable/disable tasks

This patch runs both ways.